chore: extract overview utils
